Don't display variable fields in list_products()
The current table is large and filled with NaN values.

If the purpose of the command is to view available products, it makes
more sense to only show the fixed (match) fields specific to
the product to highlight their differences.

The id field should also be hidden, as it's a postgres implementation
detail. On all UIs and APIs the unique 'name' field is effectively the
id.